Change flags to uint8

+++ test/V4Router.t.sol
@@ -101,7 +101,7 @@ contract Fallback is V4RouterTest {
 
     // Optimized router call
     (bool ok,) = payable(routerAddr).call(
-      abi.encodePacked(UNI, uint96(100_000), uint96(desc.minReturnAmount), uint256(0), data)
+      abi.encodePacked(UNI, uint96(100_000), uint96(desc.minReturnAmount), uint8(0), data)
     );
 
     assertTrue(ok, "Swap failed");
@@ -133,7 +133,7 @@ contract Fallback is V4RouterTest {
     IERC20(USDC).approve(routerAddr, 10_000_000);
     uint256 startingBalance = IERC20(USDC).balanceOf(swapSenderAddress);
     (bool ok,) = payable(routerAddr).call(
-      abi.encodePacked(UNI, uint96(10_000_000), uint96(desc.minReturnAmount), uint256(0), data)
+      abi.encodePacked(UNI, uint96(10_000_000), uint96(desc.minReturnAmount), uint8(0), data)
     );
     uint256 endingBalance = IERC20(USDC).balanceOf(swapSenderAddress);
 
@@ -146,7 +146,7 @@ contract Fallback is V4RouterTest {
     IERC20(USDC).approve(routerAddr, 250_000);
     uint256 startingBalance = IERC20(USDC).balanceOf(swapSenderAddress);
     (bool ok,) = payable(routerAddr).call(
-      abi.encodePacked(address(0), uint96(250_000), uint96(desc.minReturnAmount), uint256(0), data)
+      abi.encodePacked(address(0), uint96(250_000), uint96(desc.minReturnAmount), uint8(0), data)
     );
     uint256 endingBalance = IERC20(USDC).balanceOf(swapSenderAddress);
