+++ test/V5Router.t.sol
@@ -107,7 +107,7 @@ contract Fallback is V5RouterTest {
 
     // Optimized router call
     (bool ok,) = payable(routerAddr).call(
-      abi.encodePacked(UNI, uint96(100_000), uint96(desc.minReturnAmount), uint256(0), data)
+      abi.encodePacked(UNI, uint96(100_000), uint96(desc.minReturnAmount), uint8(0), data)
     );
 
     assertTrue(ok, "Swap failed");
@@ -139,7 +139,7 @@ contract Fallback is V5RouterTest {
     IERC20(USDC).approve(routerAddr, 10_000_000);
     uint256 startingBalance = IERC20(USDC).balanceOf(swapSenderAddress);
     (bool ok,) = payable(routerAddr).call(
-      abi.encodePacked(UNI, uint96(10_000_000), uint96(desc.minReturnAmount), uint256(0), data)
+      abi.encodePacked(UNI, uint96(10_000_000), uint96(desc.minReturnAmount), uint8(0), data)
     );
     uint256 endingBalance = IERC20(USDC).balanceOf(swapSenderAddress);
 
@@ -153,7 +153,7 @@ contract Fallback is V5RouterTest {
     uint256 startingBalance = IERC20(USDC).balanceOf(swapSenderAddress);
     (bool ok,) = payable(routerAddr).call(
       abi.encodePacked(
-        address(0), uint96(250_000), uint96(desc.minReturnAmount), uint256(0), bytes(data)
+        address(0), uint96(250_000), uint96(desc.minReturnAmount), uint8(0), bytes(data)
       )
     );
     uint256 endingBalance = IERC20(USDC).balanceOf(swapSenderAddress);
